Moved sssom_schema iports to a new `schema.py`
