+++ sssom/writers.py
@@ -16,9 +16,9 @@ from rdflib.namespace import OWL, RDF
 # from .sssom_datamodel import slots
 from sssom_schema import slots
 
+from sssom.schema import SCHEMA_YAML
 from sssom.validators import check_all_prefixes_in_curie_map
 
-from .constants import SCHEMA_YAML
 from .parsers import to_mapping_set_document
 from .util import (
     PREFIX_MAP_KEY,
